+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.0.3
+- Removed firebase for better searchability in the future
+
 ## 1.0.2
 - Added quick-links so users can visit a site easily
 - Updated generic-components